# docs/openapi/gsf-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GSF API",
    "version": "0.1.1",
    "description": "Public HTTP surface of GSF, served by the Next.js route handlers. Generated by `pnpm openapi` — do not edit by hand."
  },
  "paths": {
...
    "/api/chat/cancel": {
      "post": {
        "tags": ["chat"],
        "operationId": "post_api_chat_cancel",
        "x-gsf-upstream": "/api/chat/cancel",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "query",
            "required": true,
            "description": "Conversation whose in-flight run to abort. Must belong to the caller.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "404": {
            "description": "Missing/unknown `conversationId`, or it belongs to another user.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "const": "Conversation not found"
                }
              }
            }
          },
          "502": {
            "description": "The backend refused the cancel. The upstream status is used when it has one, 502 otherwise; the body is plain text either way.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "const": "Upstream cancel request failed"
                }
              }
            }
          }
        },
        "x-gsf-source": "frontend/app/api/chat/cancel/route.ts",
        "x-gsf-permissions": {
          "conversation": ["write"]
        },
        "x-gsf-proxied": false
      }
    },
    "/api/chat/completions": {
      "post": {
        "tags": ["chat"],
        "operationId": "post_api_chat_completions",
        "x-gsf-upstream": "/api/chat/completions",
        "requestBody": {
          "required": true,
          "description": "Step 1 of a chat turn: SQL plus a formatted answer. Charts are a separate `POST /api/chat/visualize` call. An unknown or unowned `conversation_id` is treated as absent and a new conversation is created for the caller.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "minLength": 1
                  },
                  "conversation_id": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "target_db": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "prediction": {
                    "anyOf": [
                      {
                        "type": "boolean"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": ["question"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent events; the stream ends with `data: [DONE]`"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "409": {
            "description": "Relayed from the backend: a run is already in flight for this conversation, or the semantic layer has not been compiled yet.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "required": ["detail"],
                  "additionalProperties": false
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "x-gsf-source": "frontend/app/api/chat/completions/route.ts",
        "x-gsf-permissions": {
          "chat": ["use"]
        },
        "x-gsf-proxied": false
      }
    },
    "/api/chat/visualize": {
      "post": {
        "tags": ["chat"],
        "description": "Same permission as the chat completions route: every chat user may trigger\nthis, not just admins.\n\nUnlike the completions route, this one never creates a conversation: by the\ntime step 2 runs, step 1 (or `POST /api/conversations`) has already created\nit. Falling back to create-on-missing here would leave an orphan\nconversation — just the chart/table bubble, with no question or Message 1\nahead of it — whenever this is called with an id that never got that far\n(e.g. a standalone API caller skipping the completions step).",
        "operationId": "post_api_chat_visualize",
        "x-gsf-upstream": "/api/chat/visualize",
        "requestBody": {
          "required": false,
          "description": "Step 2 of chat: the SQL and its already-executed result from step 1. A missing or unparseable body is tolerated — it degrades to `{ \"charts\": null }` rather than erroring, because the chart is a nice-to-have.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "minLength": 1
                  },
                  "sql": {
                    "default": "",
                    "type": "string"
                  },
                  "result": {
                    "description": "The executed SQL result — the `sql_response_from_db` from the step 1 answer. Either a one-item list holding a JSON-records string, or a list of row objects."
                  },
                  "conversation_id": {
                    "description": "When given and owned by the caller, the result and any chart are appended to that conversation as an assistant message. Ignored by the backend; consumed by this handler only.",
                    "type": "string"
                  }
                },
                "required": ["question", "result"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "x-gsf-source": "frontend/app/api/chat/visualize/route.ts",
        "x-gsf-permissions": {
          "chat": ["use"]
        },
        "x-gsf-proxied": false
      }
    },
    "/api/chat/watch": {
      "get": {
        "tags": ["chat"],
        "operationId": "get_api_chat_watch",
        "x-gsf-upstream": "/api/chat/watch",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "query",
            "required": true,
            "description": "Conversation whose in-flight run to observe. Must belong to the caller.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent events; the stream ends with `data: [DONE]`"
          },
          "400": {
            "description": "No `conversation_id` was given.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "const": "Missing conversation_id"
                }
              }
            }
          },
          "401": {
            "description": "No authenticated caller could be resolved"
          },
          "403": {
            "description": "The caller lacks the required permission"
          },
          "404": {
            "description": "Unknown conversation, or it belongs to another user.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "const": "Conversation not found"
                }
              }
            }
          },
          "502": {
            "description": "The backend refused the watch. The upstream status is used when it has one, 502 otherwise; the body is plain text either way.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "const": "Upstream watch request failed"
                }
              }
            }
          }
        },
        "x-gsf-source": "frontend/app/api/chat/watch/route.ts",
        "x-gsf-permissions": {
          "conversation": ["read"]
        },
        "x-gsf-proxied": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChartsResponse": {
        "additionalProperties": true,
        "description": "``charts`` is null when the model declined to propose a chart.",
        "properties": {
          "charts": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Charts"
          }
        },
        "required": ["charts"],
        "title": "ChartsResponse",
        "type": "object"
      },
      "ChatCancelResponse": {
        "additionalProperties": true,
        "properties": {
          "cancelled": {
            "title": "Cancelled",
            "type": "boolean"
          }
        },
        "required": ["cancelled"],
        "title": "ChatCancelResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  }
}